Move RelativeTime and UserAvatar below the "OpenProject" folder
