+++ lookbook/previews/open_project/users/avatar_component_preview/sizes.html.erb
